+++ src/views/HomeView.vue
@@ -1,15 +1,15 @@
 <template>
 	<v-container fluid class=''>
 		<v-row align='center' justify='center'>
-			<v-col cols='10' md='8' align='center' justify='center'>
-				<p class='text-center text-white text-h5'>
-					Since May 9th 2015, two friends have pedantically exchanged information on every evening meal that they have consumed.
+			<v-col cols='11' align='center' >
+				<p class='text-white text-h5'>
+					Since May 9th 2015, two friends have fastidiously exchanged information on every evening meal that they have consumed.
 				</p>
 				<p class='text-center text-white text-h5 my-4'>
 					This is a comprehensive chronicling of that pedantry.
 				</p>
 			</v-col>
-			<v-col cols='10' align='center' justify='center'>
+			<v-col cols='11' align='center' justify='center'>
 				<v-btn :to='FrontEndRoutes.MEALS' color='secondary' class='cl' size='x-large' variant='flat' rounded>
 					<v-row align='center' justify='center'>
 						<v-col cols='auto' class='pa-1'>