+++ CHANGELOG.md
@@ -1,3 +1,32 @@
+# <a href='https://github.com/mrjackwills/mealpedant_vue/releases/tag/v2.0.1'>v2.0.1</a>
+### 2025-05-23
+
+### Chores
++ dependencies updated, [dc1e0ed7](https://github.com/mrjackwills/mealpedant_vue/commit/dc1e0ed7858170a6e39230f039e98ee529f9fe94), [70670ba1](https://github.com/mrjackwills/mealpedant_vue/commit/70670ba135f767363297dcec9d5a6e28e33db88c), [816af6f3](https://github.com/mrjackwills/mealpedant_vue/commit/816af6f3324dbbb5bb947ab46794c3759e573ed1), [81743294](https://github.com/mrjackwills/mealpedant_vue/commit/817432949ff20e86d26601e41b51422c74b64f5a), [a78b71b0](https://github.com/mrjackwills/mealpedant_vue/commit/a78b71b0a1e498fd8a24736b45569a5f17a9ac44),
+
+### Features
++ category chip flat if selected, [6088fae8](https://github.com/mrjackwills/mealpedant_vue/commit/6088fae8745f411e23c0240fa24f6a990ac8598f)
++ mono text class, [ba4e21ca](https://github.com/mrjackwills/mealpedant_vue/commit/ba4e21cac1f1388e7798892937014d120ee8d00a)
+
+### Fixes
++ add/edit meal category hint size, [d53e4a62](https://github.com/mrjackwills/mealpedant_vue/commit/d53e4a621a0a7a0ea990fd62ddf860e41bed392a), [f1cd7b77](https://github.com/mrjackwills/mealpedant_vue/commit/f1cd7b774db19cb48bda543c2c44d143affa7fc5)
++ admin photo row/cols, [7c095fa1](https://github.com/mrjackwills/mealpedant_vue/commit/7c095fa1bb940e51b30ae50946d59c38d9e41d4f), [8bc3179e](https://github.com/mrjackwills/mealpedant_vue/commit/8bc3179ee1a20da7090cbe51ad348572721ad4be)
++ admin table mobile size, [4147dd38](https://github.com/mrjackwills/mealpedant_vue/commit/4147dd384aed1f18145f59e3dd188cf16f1505cb)
++ category spacing on mobile, [4f830271](https://github.com/mrjackwills/mealpedant_vue/commit/4f830271b3ad075a0570b4347bf344aba551bb58)
++ change description, [a0414a9e](https://github.com/mrjackwills/mealpedant_vue/commit/a0414a9e40b838a5ee3bc7060de6242bbadce5b5), [c88061de](https://github.com/mrjackwills/mealpedant_vue/commit/c88061dee337434614fe8a777aab59ededbd9aad)
++ create backup negative margin, [c4948a77](https://github.com/mrjackwills/mealpedant_vue/commit/c4948a778d0e034b536d335583732b9b8e734a9e)
++ create_release.sh, [748b5813](https://github.com/mrjackwills/mealpedant_vue/commit/748b581334ed4b651378536eb70310a61efe2a10)
++ hardcode github url in page, [c38e3a41](https://github.com/mrjackwills/mealpedant_vue/commit/c38e3a41da157629f2a3550486e57ea5c496d72d)
++ homeview width, [4449418c](https://github.com/mrjackwills/mealpedant_vue/commit/4449418c3b140c588ca2ae3a54c7612355231e59)
++ mealview layout, [00afd178](https://github.com/mrjackwills/mealpedant_vue/commit/00afd1789f8dc7c69039fdea5070db3ac6e0fb93)
++ snackbar layout & size, [e570fb0c](https://github.com/mrjackwills/mealpedant_vue/commit/e570fb0c20057dedf54a1f2dcc0b19d990196eb6), [fb400091](https://github.com/mrjackwills/mealpedant_vue/commit/fb400091b159f9b20db057faa84ec6a0b5f7c0ff)
++ unused template removed, [d0fe03ee](https://github.com/mrjackwills/mealpedant_vue/commit/d0fe03ee2c857949a62752a25650792bbdd78795)
++ unused justify removed, [f5b1a6e0](https://github.com/mrjackwills/mealpedant_vue/commit/f5b1a6e02fafeebbe1be9a0309a0d518b4ca1778)
++ trim description, original photo link, [89a8fae4](https://github.com/mrjackwills/mealpedant_vue/commit/89a8fae4f9f876d4a21273e02ad77c6a90b48afc)
+
+### Refactors
++ remove errorsnack from mealStorage, [8fdf8d97](https://github.com/mrjackwills/mealpedant_vue/commit/8fdf8d978e6e23e81598ee786df9aba9292df69b)
+
 # <a href='https://github.com/mrjackwills/mealpedant_vue/releases/tag/v2.0.0'>v2.0.0</a>
 ### 2025-04-06